refactor(files):重构项目文件系统结构 1. 创建了清晰的模块化结构：       - easyki/ 核心转换引擎       - ui/web/ Web界面组件       - scripts/ 启动脚本
   2. 更新了所有导入语句以匹配新结构

   3. 修正了相关文档以反映新的目录组织

   4. 验证了所有功能正常运行
