+++ easyki/easyeda/easyeda_importer.py
@@ -2,8 +2,8 @@
 import json
 import logging
 
-from easyeda.easyeda_api import EasyedaApi
-from easyeda.parameters_easyeda import *
+from easyki.easyeda.easyeda_api import EasyedaApi
+from easyki.easyeda.parameters_easyeda import *
 
 
 def add_easyeda_pin(pin_data: str, ee_symbol: EeSymbol):
