+++ easyki/easyeda/parameters_easyeda.py
@@ -5,7 +5,7 @@ from typing import List, Union
 
 from pydantic import BaseModel, field_validator
 
-from easyeda.svg_path_parser import parse_svg_path
+from easyki.easyeda.svg_path_parser import parse_svg_path
 
 
 class EasyedaPinType(Enum):
