+++ easyki/easyeda/svg_path_parser.py
