+++ easyki/kicad/export_kicad_3d_model.py
@@ -8,8 +8,8 @@ from __future__ import annotations
 import re
 import textwrap
 
-from easyeda.parameters_easyeda import Ee3dModel
-from kicad.parameters_kicad_footprint import Ki3dModel
+from easyki.easyeda.parameters_easyeda import Ee3dModel
+from easyki.kicad.parameters_kicad_footprint import Ki3dModel
 
 VRML_HEADER = """#VRML V2.0 utf8
 # 3D model generated by easyeda2kicad.py"""
