+++ easyki/kicad/export_kicad_footprint.py
@@ -3,8 +3,8 @@ import logging
 from math import acos, cos, isnan, pi, sin, sqrt
 from typing import Tuple, Union
 
-from easyeda.parameters_easyeda import ee_footprint
-from kicad.parameters_kicad_footprint import *
+from easyki.easyeda.parameters_easyeda import ee_footprint
+from easyki.kicad.parameters_kicad_footprint import *
 
 # ---------------------------------------
