+++ easyki/kicad/parameters_kicad_symbol.py
